# CP-54706: Send pause command to xenguest earlier