When emu-manager decides that xenguest has done enough in the live phase, it sets the state to live_done and then proceeds to suspend the VM. While the VM is executing its suspend process, xenguest continues to copy dirtied pages and this can slow down the suspend process.

In cases where vgpu is not used, it is preferable to send the pause command as soon as the live_done state is reached. This prevents excessive iterations from xenguest, lets the VM's suspend process run quickly, and the overall migration finish faster.

Measurements of the time from "xenguest live stage is done" to "Finished, send complete":

Before: 1020 ms, 1284 ms, 1263 ms
After: 439 ms, 470 ms, 431 ms