@TSnake41 2025-10-02
That seems contradictory to xenguest expectations
https://github.com/xenserver/xen.pg/blob/XS-8.4/patches/xenguest.patch#L125-L127

I am not sure what is actually correct; but I suppose xenopsd needs to acknowledge live migration before starting to actually live migrate for instance to make PV backends consistent (i.e avoiding them to write to guest memory while we actually live migrate).